add support for running in cargo workspaces
This adds cargo workspace support. When `cargo add-dynamic` is invoked, we will
try to find a workspace toml file starting from cwd. We will check if the target
package (the package the dylib should be added to) is a workspace member and if
this is the case we will add the dylib package as a member as well.

fixes #4